(Wars) Level ID Map
Not too sure if I should list the credits stages (such as w5s01), as they aren't actually playable stages and are just there to load a specific GEdit for the credits camera flybys.

Might also want to split up the stage list to three different ones for Main, Secret and Extra?